+++ src/argument_processors/argument_processor_database.h
@@ -22,4 +22,10 @@ class ArgumentProcessorDatabaseStruct : public WsjcppArgumentProcessor {
         virtual int exec(const std::vector<std::string> &vRoutes, const std::vector<std::string> &vSubParams) override;
 };
 
+class ArgumentProcessorDatabaseCreate : public WsjcppArgumentProcessor {
+    public:
+        ArgumentProcessorDatabaseCreate();
+        virtual int exec(const std::vector<std::string> &vRoutes, const std::vector<std::string> &vSubParams) override;
+};
+
 #endif // ARGUMENT_PROCESSOR_DATABASE_H